Conflicts:
	TangoHapps.xcodeproj/project.pbxproj
	TangoHapps.xcodeproj/project.xcworkspace/xcshareddata/TangoHapps.xccheckout
	TangoHapps.xcodeproj/project.xcworkspace/xcuserdata/avenix.xcuserdatad/UserInterfaceState.xcuserstate
	TangoHapps.xcodeproj/xcuserdata/avenix.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
	TangoHapps/classes/objects/layers/THEditor.m
	TangoHapps/classes/palette properties/THGestureProperties.h
	TangoHapps/classes/palette properties/THGestureProperties.m
	TangoHapps/resources/gui/properties/THGestureProperties.xib
	TangoHapps/resources/images/project/palette/gesture.png
	TangoHapps/resources/images/project/palette/gesture@2x.png
	TangoHapps/resources/images/project/palette/palette_gesture.png
	TangoHapps/resources/images/project/palette/palette_gesture@2x.png

+++ Frameworks/libpd/pure-data/src/d_array.c
@@ -233,8 +233,8 @@ static void tabplay_tilde_list(t_tabplay_tilde *x, t_symbol *s,
     if (length <= 0)
         x->x_limit = 0x7fffffff;
     else
-        x->x_limit = start + length;
-    x->x_phase = start;
+        x->x_limit = (int) (start + length);
+    x->x_phase = (int)start;
 }
 
 static void tabplay_tilde_stop(t_tabplay_tilde *x)
@@ -716,7 +716,7 @@ static t_int *tabsend_perform(t_int *w)
 {
     t_tabsend *x = (t_tabsend *)(w[1]);
     t_sample *in = (t_sample *)(w[2]);
-    int n = w[3];
+    int n = (int) w[3];
     t_word *dest = x->x_vec;
     int i = x->x_graphcount;
     if (!x->x_vec) goto bad;
@@ -790,7 +790,7 @@ static t_int *tabreceive_perform(t_int *w)
 {
     t_tabreceive *x = (t_tabreceive *)(w[1]);
     t_sample *out = (t_sample *)(w[2]);
-    int n = w[3];
+    int n = (int) w[3];
     t_word *from = x->x_vec;
     if (from)
     {
